+++ account-decoder/Cargo.toml
@@ -21,7 +21,6 @@ serde_derive = "1.0.103"
 serde_json = "1.0.56"
 solana-config-program = { path = "../programs/config", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-vote-program = { path = "../programs/vote", version = "=1.7.2" }
 spl-token-v2-0 = { package = "spl-token", version = "=3.1.1", features = ["no-entrypoint"] }
 thiserror = "1.0"
