+++ cli-output/Cargo.toml
@@ -23,7 +23,6 @@ solana-account-decoder = { path = "../account-decoder", version = "=1.7.2" }
 solana-clap-utils = { path = "../clap-utils", version = "=1.7.2" }
 solana-client = { path = "../client", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.7.2" }
 solana-vote-program = { path = "../programs/vote", version = "=1.7.2" }
 spl-memo = { version = "=3.0.1", features = ["no-entrypoint"] }
