+++ cli/Cargo.toml
@@ -41,7 +41,6 @@ solana-net-utils = { path = "../net-utils", version = "=1.7.2" }
 solana_rbpf = "=0.2.11"
 solana-remote-wallet = { path = "../remote-wallet", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.7.2" }
 solana-version = { path = "../version", version = "=1.7.2" }
 solana-vote-program = { path = "../programs/vote", version = "=1.7.2" }
