+++ core/Cargo.toml
@@ -62,7 +62,6 @@ solana-runtime = { path = "../runtime", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
 solana-frozen-abi = { path = "../frozen-abi", version = "=1.7.2" }
 solana-frozen-abi-macro = { path = "../frozen-abi/macro", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-streamer = { path = "../streamer", version = "=1.7.2" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.7.2" }
 solana-version = { path = "../version", version = "=1.7.2" }
@@ -81,6 +80,7 @@ num_cpus = "1.13.0"
 reqwest = { version = "0.11.2", default-features = false, features = ["blocking", "rustls-tls", "json"] }
 serde_json = "1.0.56"
 serial_test = "0.4.0"
+solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-version = { path = "../version", version = "=1.7.2" }
 symlink = "0.1.0"
 systemstat = "0.1.5"
