+++ ledger/Cargo.toml
@@ -46,7 +46,6 @@ solana-perf = { path = "../perf", version = "=1.7.2" }
 solana-rayon-threadlimit = { path = "../rayon-threadlimit", version = "=1.7.2" }
 solana-runtime = { path = "../runtime", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-storage-bigtable = { path = "../storage-bigtable", version = "=1.7.2" }
 solana-storage-proto = { path = "../storage-proto", version = "=1.7.2" }
 solana-vote-program = { path = "../programs/vote", version = "=1.7.2" }
@@ -67,7 +66,6 @@ features = ["lz4"]
 assert_matches = "1.3.0"
 matches = "0.1.6"
 solana-account-decoder = { path = "../account-decoder", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 
 [build-dependencies]
 rustc_version = "0.2"
