+++ perf/Cargo.toml
@@ -23,7 +23,6 @@ solana-logger = { path = "../logger", version = "=1.7.2" }
 solana-metrics = { path = "../metrics", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
 solana-rayon-threadlimit = { path = "../rayon-threadlimit", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 
 [lib]
 name = "solana_perf"
