+++ perf/src/test_tx.rs
@@ -1,10 +1,12 @@
-use solana_sdk::hash::Hash;
-use solana_sdk::instruction::CompiledInstruction;
-use solana_sdk::signature::{Keypair, Signer};
-use solana_sdk::system_instruction::SystemInstruction;
-use solana_sdk::system_program;
-use solana_sdk::system_transaction;
-use solana_sdk::transaction::Transaction;
+use solana_sdk::{
+    hash::Hash,
+    instruction::CompiledInstruction,
+    signature::{Keypair, Signer},
+    stake,
+    system_instruction::SystemInstruction,
+    system_program, system_transaction,
+    transaction::Transaction,
+};
 
 pub fn test_tx() -> Transaction {
     let keypair1 = Keypair::new();
@@ -22,7 +24,7 @@ pub fn test_multisig_tx() -> Transaction {
 
     let transfer_instruction = SystemInstruction::Transfer { lamports };
 
-    let program_ids = vec![system_program::id(), solana_stake_program::id()];
+    let program_ids = vec![system_program::id(), stake::program::id()];
 
     let instructions = vec![CompiledInstruction::new(
         0,
