+++ program-test/Cargo.toml
@@ -29,4 +29,3 @@ tokio = { version = "1", features = ["full"] }
 
 [dev-dependencies]
 assert_matches = "1.3.0"
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
