+++ programs/stake/src/lib.rs
@@ -2,15 +2,16 @@
 #![allow(clippy::integer_arithmetic)]
 use solana_sdk::genesis_config::GenesisConfig;
 
+#[deprecated(
+    since = "1.7.2",
+    note = "Please use `solana_sdk::stake::program::id` or `solana_program::stake::program::id` instead"
+)]
+pub use solana_sdk::stake::program::{check_id, id};
+
 pub mod config;
 pub mod stake_instruction;
 pub mod stake_state;
 
-solana_sdk::declare_id!("Stake11111111111111111111111111111111111111");
-
 pub fn add_genesis_accounts(genesis_config: &mut GenesisConfig) -> u64 {
     config::add_genesis_account(genesis_config)
 }
-
-#[macro_use]
-extern crate solana_frozen_abi_macro;
