+++ rpc/Cargo.toml
@@ -38,7 +38,6 @@ solana-perf = { path = "../perf", version = "=1.7.2" }
 solana-poh = { path = "../poh", version = "=1.7.2" }
 solana-runtime = { path = "../runtime", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-storage-bigtable = { path = "../storage-bigtable", version = "=1.7.2" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.7.2" }
 solana-version = { path = "../version", version = "=1.7.2" }
@@ -52,6 +51,7 @@ tokio-util = { version = "0.3", features = ["codec"] } # This crate needs to sta
 serial_test = "0.4.0"
 solana-logger = { path = "../logger", version = "=1.7.2" }
 solana-net-utils = { path = "../net-utils", version = "=1.7.2" }
+solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 symlink = "0.1.0"
 
 [lib]
