+++ runtime/src/genesis_utils.rs
@@ -8,10 +8,10 @@ use solana_sdk::{
     pubkey::Pubkey,
     rent::Rent,
     signature::{Keypair, Signer},
+    stake::state::StakeState,
     system_program,
 };
 use solana_stake_program::stake_state;
-use solana_stake_program::stake_state::StakeState;
 use solana_vote_program::vote_state;
 use std::borrow::Borrow;
