+++ sdk/program/src/stake/mod.rs
@@ -1,0 +1,7 @@
+pub mod config;
+pub mod instruction;
+pub mod state;
+
+pub mod program {
+    crate::declare_id!("Stake11111111111111111111111111111111111111");
+}
