+++ stake-accounts/src/main.rs
@@ -16,12 +16,10 @@ use solana_sdk::{
     pubkey::Pubkey,
     signature::{unique_signers, Signature, Signer},
     signers::Signers,
+    stake::{instruction::LockupArgs, state::Lockup},
     transaction::Transaction,
 };
-use solana_stake_program::{
-    stake_instruction::LockupArgs,
-    stake_state::{Lockup, StakeState},
-};
+use solana_stake_program::stake_state;
 use std::env;
 use std::error::Error;
 
@@ -52,7 +50,7 @@ fn get_balances(
 fn get_lockup(client: &RpcClient, address: &Pubkey) -> Result<Lockup, ClientError> {
     client
         .get_account(address)
-        .map(|account| StakeState::lockup_from(&account).unwrap())
+        .map(|account| stake_state::lockup_from(&account).unwrap())
 }
 
 fn get_lockups(
