+++ stake-monitor/Cargo.toml
@@ -22,7 +22,6 @@ solana-logger = { path = "../logger", version = "=1.7.2" }
 solana-metrics = { path = "../metrics", version = "=1.7.2" }
 solana-rpc = { path = "../rpc", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.7.2" }
 solana-version = { path = "../version", version = "=1.7.2" }
