+++ tokens/Cargo.toml
@@ -27,7 +27,6 @@ solana-client = { path = "../client", version = "=1.7.2" }
 solana-remote-wallet = { path = "../remote-wallet", version = "=1.7.2" }
 solana-runtime = { path = "../runtime", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.7.2" }
 solana-version = { path = "../version", version = "=1.7.2" }
 spl-associated-token-account-v1-0 = { package = "spl-associated-token-account", version = "=1.0.2" }
