+++ transaction-status/Cargo.toml
@@ -21,7 +21,6 @@ serde_json = "1.0.56"
 solana-account-decoder = { path = "../account-decoder", version = "=1.7.2" }
 solana-sdk = { path = "../sdk", version = "=1.7.2" }
 solana-runtime = { path = "../runtime", version = "=1.7.2" }
-solana-stake-program = { path = "../programs/stake", version = "=1.7.2" }
 solana-vote-program = { path = "../programs/vote", version = "=1.7.2" }
 spl-associated-token-account-v1-0 = { package = "spl-associated-token-account", version = "=1.0.2", features = ["no-entrypoint"] }
 spl-memo = { version = "=3.0.1", features = ["no-entrypoint"] }
